fix how to get activated tab
state.url return the current url as /logs or /logs?queryParams=... When you are looking for tab.url == state.url with query params. It's always false.

I used the first child of the application which represents the navigation tab. The parent is the default empty path. Well, the children could be null (ex: cc.ilucca.net/), it's so I added a "?".