fix disabled state cannot work with the usecase
Why ?
With the disabled state, it's not possible to handle formArray state (validators). In the edition mode even if the grid is disabled, we need to know her state, we need to refresh validator when we add a none disabled row..

That's why, I created the readonly mode. 
HTML5 used this field but not Angular. https://developer.mozilla.org/fr/docs/Web/HTML/Attributes/readonly